+++ Modules/Intent.Modules.Azure.BlobStorage/Templates/AzureBlobStorageImplementation/AzureBlobStorageImplementationTemplatePartial.cs
@@ -30,15 +30,23 @@ namespace Intent.Modules.Azure.BlobStorage.Templates.AzureBlobStorageImplementat
 
         public override void BeforeTemplateExecution()
         {
-            var useManagedIdentity = ExecutionContext.Settings.GetBlobStorageSettings().AuthenticationMethods()?.Any(x => x.IsManagedIdentity()) ?? false;
+            var authMethods = ExecutionContext.Settings.GetBlobStorageSettings().AuthenticationMethods() ?? [];
+            var useManagedIdentity = authMethods.Any(x => x.IsManagedIdentity());
+            var useKeyBased = !useManagedIdentity || authMethods.Any(x => x.IsKeyBased());
 
             if (useManagedIdentity)
             {
                 ExecutionContext.EventDispatcher.Publish(new AppSettingRegistrationRequest("AzureBlobStorageServiceUri", "https://<storage-account>.blob.core.windows.net"));
             }
-            else
+
+            if (useKeyBased)
             {
                 ExecutionContext.EventDispatcher.Publish(new AppSettingRegistrationRequest("AzureBlobStorage", "UseDevelopmentStorage=true"));
+            }
+
+            if (useManagedIdentity && useKeyBased)
+            {
+                ExecutionContext.EventDispatcher.Publish(new AppSettingRegistrationRequest("AzureBlobStorageAuthenticationMethod", "key-based"));
             }
 
             ExecutionContext.EventDispatcher.Publish(ContainerRegistrationRequest.ToRegister(this)